Made the "X is a cheater" message work online.
When activating a cheat online, the client makes the changes locally, but ALSO sends the message on the server,
so that it won't waste a packet.

This behaviour isn't the best, especially since we have notarget that will desync from a client to another.

Besides, the "give" command won't be sent correctly to the server. This will be done in a future commit.